split out the db cli
