+++ api/models.py
@@ -1,5 +1,3 @@
-import click
-from flask.cli import with_appcontext
 from flask_sqlalchemy import SQLAlchemy
 
 db = SQLAlchemy()
@@ -38,64 +36,11 @@ class Matrix(db.Model):
     __tablename__ = "matrices"
 
     id = db.Column(db.Integer, primary_key=True)
-    user_id = db.Column(db.Integer, db.ForeignKey("users.id"))
+    user_id = db.Column(db.Integer, db.ForeignKey("users.id"), nullable=False)
     answer_bottom_right = db.Column(db.String(255), nullable=False)
     answer_bottom_left = db.Column(db.String(255), nullable=False)
     answer_top_left = db.Column(db.String(255), nullable=False)
     answer_top_right = db.Column(db.String(255), nullable=False)
     created_on = db.Column(db.DateTime(timezone=True), server_default=db.func.now())
 
-
-@click.command("db-create")
-@with_appcontext
-def db_create_cmd():
-    db.create_all()
-    click.echo("Initialized the database.")
-
-
-@click.command("db-drop")
-@with_appcontext
-def db_drop_cmd():
-    db.drop_all()
-    click.echo("Dropped the database.")
-
-
-@click.command("db-seed")
-@with_appcontext
-def db_seed_cmd():
-    user1 = User(
-        first_name="Joe",
-        last_name="Wolfe",
-        username="joe.wolfe",
-        email="joe.wolfe@gmail.com",
-        password="pass",
-    )
-    user2 = User(
-        first_name="Sam",
-        last_name="Wolfe",
-        username="sam.wolfe",
-        email="sam.wolfe@gmail.com",
-        password="pass",
-    )
-    user3 = User(
-        first_name="Tina",
-        last_name="Long",
-        username="tina.long",
-        email="tina.long@gmail.com",
-        password="pass",
-    )
-    user4 = User(
-        first_name="Cory",
-        last_name="Long",
-        username="cory.long",
-        email="cory.long@gmail.com",
-        password="pass",
-    )
-
-    db.session.add(user1)
-    db.session.add(user2)
-    db.session.add(user3)
-    db.session.add(user4)
-
-    db.session.commit()
-    click.echo("Seeded the database.")
+    user = db.relationship("User")
